Random sample of 300 rare species.
Assessments of rarity by taxonomic specialists at the Missouri Botanical Garden and the New York Botanical Garden for a random sample of 300 species with three observations or fewer in the BIEN database. Most species (72.7%) identified as “rare” based on the number of unique occurrences within the BIEN database are also recognized as rare by experts.